+++ disparity_view/zed_capture.py
@@ -17,8 +17,8 @@ from pathlib import Path
 import cv2
 import numpy as np
 
-from disparityview.disparity_view import as_colorimage, get_dirs
-from disparityview.zed_camerainfo import CameraParameter
+from disparity_view.disparity_view import as_colorimage, get_dirs
+from disparity_view.zed_camerainfo import CameraParameter
 
 MAX_ABS_DEPTH, MIN_ABS_DEPTH = 0.0, 2.0  # [m]
